feat(client): tentatively finish AdminView (#12)

+++ server/routes/admin.mts
@@ -1,11 +1,22 @@
 import type { User, CreateUserInner, CreateDateInner, DatesRecord, CreateSocietyInner, Society } from "../../types/types.js";
 import RequestHandler from "../services/request-handler.mjs";
 
+class AdminUserGetHandler extends RequestHandler {
+    static method = RequestHandler.GET;
+    static path = "/user/list";
+
+    public async handle_core(): Promise<User[]> {
+        await this.authorize();
+        return await this.databaseService.list_users();
+    }
+}
+
 class AdminUserImportHandler extends RequestHandler {
     static method = RequestHandler.POST;
     static path = "/user/import";
 
     public async handle_core(): Promise<User[]> {
+        await this.authorize();
         const users = this.req.body as CreateUserInner[];
         return await Promise.all(users.map(user => this.databaseService.create_or_update_user(user)));
     }
@@ -16,9 +27,20 @@ class AdminUserDeleteHandler extends RequestHandler {
     static path = "/user/delete";
 
     public async handle_core(): Promise<{ success: true }> {
+        await this.authorize();
         const user_ids = this.req.body as string[];
         await Promise.all(user_ids.map(user_id => this.databaseService.delete_user(user_id)));
         return { success: true };
+    }
+}
+
+class AdminDateGetHandler extends RequestHandler {
+    static method = RequestHandler.GET;
+    static path = "/date/list";
+
+    public async handle_core(): Promise<DatesRecord[]> {
+        await this.authorize();
+        return await this.databaseService.list_dates();
     }
 }
 
@@ -27,6 +49,7 @@ class AdminDateCreateHandler extends RequestHandler {
     static path = "/date/create";
 
     public async handle_core(): Promise<DatesRecord> {
+        await this.authorize();
         const date = this.req.body as CreateDateInner;
         return await this.databaseService.create_date(date);
     }
@@ -37,6 +60,7 @@ class AdminDateActivateHandler extends RequestHandler {
     static path = "/date/activate";
 
     public async handle_core(): Promise<DatesRecord> {
+        await this.authorize();
         const { id, activate } = this.req.body as { id: string, activate: boolean };
         return await this.databaseService.toggle_activate_date(id, activate);
     }
@@ -47,6 +71,7 @@ class AdminSocietyImportHandler extends RequestHandler {
     static path = "/society/import";
 
     public async handle_core(): Promise<Society[]> {
+        await this.authorize();
         const societies = this.req.body as CreateSocietyInner[];
         return await Promise.all(societies.map(society => this.databaseService.create_or_update_society(society)));
     }
@@ -57,6 +82,7 @@ class AdminSocietyDeleteHandler extends RequestHandler {
     static path = "/society/delete";
 
     public async handle_core(): Promise<{ success: true }> {
+        await this.authorize();
         const society_ids = this.req.body as string[];
         await Promise.all(society_ids.map(society_id => this.databaseService.delete_society(society_id)));
         return { success: true };
@@ -68,13 +94,14 @@ class AdminSocietyUpdateHandler extends RequestHandler {
     static path = "/society/update";
 
     public async handle_core(): Promise<Society> {
+        await this.authorize();
         const { id, ...society } = this.req.body as Partial<CreateSocietyInner> & { id: string };
         return await this.databaseService.update_society(id, society);
     }
 }
 
 export default RequestHandler.inject(
-    AdminUserImportHandler, AdminUserDeleteHandler,
-    AdminDateCreateHandler, AdminDateActivateHandler,
+    AdminUserGetHandler, AdminUserImportHandler, AdminUserDeleteHandler,
+    AdminDateGetHandler, AdminDateCreateHandler, AdminDateActivateHandler,
     AdminSocietyImportHandler, AdminSocietyDeleteHandler, AdminSocietyUpdateHandler
 );
